Allow a service to define a prefix for its id
Support the annotation `id-prefix` to allow a service to define a prefix
it would like added to its generated kube-service-exporter id.

Materially, this id allows a service to control its sort order among
other services. We use the service id to name haproxy acls, which are
applied in the order they appear in its configuration. Therefore this
change allows us to control the order acls are processed, irrespective
of the generated kube-service-exporter id

github/compute-platform#43